net: l2: ieee802154: TSCH slot operation thread
This change introduces a TSCH thread that does not yet do much except
for logging that it is alive. The thread can be started, stopped and
queried via net_mgmt commands and sets appropriate status flags in the
ieee802154_context object.

This implements the basic infrastructure for the MLME-TSCH-MODE request
and confirmation (see IEEE 802.15.4-2020, sections 8.2.19.5 and
8.2.19.6).

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>