[wip] treewide: introduce new settings parameter
In order to keep the user-facing API of the zephyr build as simple as
possible and re-use existing knowledge in the community, we introduce
new environment variables SETTINGS_FILES, SETTINGS_OVERLAY_FILES and
EXTRA_SETTINGS_OVERLAY_FILES.

These variables may be used to specify both, DTS and config YAML files.

It is recommended to deprecate the corresponding DTS_... variables
although they should be kept for backwards compatibility for as long as
necessary.

Signed-off-by: Florian Grandel <fgrandel@code-for-humans.de>